Added files from streaming project example
This is from the Microphones_Straming example which comes with the STM32CubeExpansion_MEMSMIC1_V3.0.0 code example package.
This code has not yet been altered or merged.

Note there are also other files which have not been copied yet as they may just be duplicates. Will need to go back and review these if there is a problem.